+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added a flexible search filter UI for the job list. [PR #XXX](https://github.com/riverqueue/riverui/pull/XXX).
+- Added a flexible search filter UI for the job list. [PR #344](https://github.com/riverqueue/riverui/pull/344).
 
 ## [v0.9.0] - 2025-04-08